requirements.txt
black==24.10.0
isort==5.13.2
requests==2.32.3
python-dotenv==1.0.1
reflex==0.6.5
openai==1.54.4
